pyproject.toml
[project]
name = "leader-teleop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "blessed>=1.21.0",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pycairo>=1.28.0",
    "pygobject==3.50.0",
    "pyyaml>=6.0.2",
    "pyzed>=1.3.0",
    "serial>=0.0.97",
]
